backend/talkeasy/requirement.txt
django
djangorestframework
django-cors-headers
django-rest-knox
djangorestframework-jwt
